refactor(gui): stage 1 — extract bridge, state, dom, selectors modules
Verbatim moves out of main.js (review with --color-moved):
- src/bridge.js: single re-export point for the Wails surface. Sits as
  a sibling of main.js so the vite substitution plugin's EXACT-literal
  specifier match ('../wailsjs/...') keeps resolving for both test
  harnesses. Explicit named re-exports (export * would make names
  ambiguous under substitution, where both specifiers resolve to one
  module).
- src/app/state.js: the shared state object + loadSavedView /
  loadSavedCollapsed / saveCollapsed + the __hive_state e2e affordance
  (VITE_WAILS_MOCK/REAL gate unchanged).
- src/app/dom.js: termsHost/projectsUL/status handles + the status
  controller (setStatus, flashStatus, reportFailure).
- src/app/selectors.js: orderedSessions, activeCwd, activeProjectId,
  resolveSessionCwd.

main.js: 3559 → 3420 lines. Only import/export lines changed; no
behavior. Suites: vitest 149, mock e2e 52(+1 skip), e2e-real 5.
